+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Fixed not being able to change some configs in blocks
 - Fixed serverside configs not syncing up with clientside
 - Fixed not being able to move inventory items in Grid GUI's to hotbar via the number keys
+- Fixed Relays when being in "Ignore Redstone" mode using up energy
 - Improved collisions of Cable parts
 
 **Features**